feat: restructure navigation components and improve loading indicators
- Moved RingLoader component to the navigation directory for better organization.
- Added Breadcrumbs component for improved navigation experience.
- Introduced SidebarContent component to manage sidebar navigation items.
- Created ProfileDropdown and Username components for user profile management.
- Implemented ProfileForm component for user profile editing.
- Added Navbar component for presentation layout with responsive design.
- Updated Main layout to include new components and improve loading states.
- Removed unused modal dialog store and related tests.
- Refactored SelectOrganization component to use a new Select component for better UI.
- Updated imports in various components to reflect new file structure.
