+++ apps/client/src/components/admin/ChangeRoleDialog.vue
@@ -7,7 +7,7 @@ import Select from "primevue/select";
 import { computed, ref } from "vue";
 import { useI18n } from "vue-i18n";
 import apiClient from "../../lib/api-client.ts";
-import RingLoader from "../RingLoader.vue";
+import RingLoader from "../navigation/RingLoader.vue";
 
 const props = defineProps<{
   userId: string;
