+++ apps/client/src/components/presentation/Navbar.vue
