+++ apps/client/src/view/Profile.vue
@@ -1,5 +1,5 @@
 <script lang="ts" setup>
-import ProfileForm from "../components/ProfileForm.vue";
+import ProfileForm from "../components/profile/ProfileForm.vue";
 </script>
 
 <template>
